requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0  # 生产环境 Web 服务器